Move sorted key list tests over to sep module
