+++ paramtools/tests/test_utils.py
@@ -9,7 +9,6 @@ from paramtools import (
     hashable_value_object,
     filter_labels,
     make_label_str,
-    SortedKeyList,
     read_json,
 )
 
@@ -147,54 +146,3 @@ def test_make_label_str():
     assert make_label_str({"value": 0}) == ""
     assert make_label_str({}) == ""
     assert make_label_str({"b": 0, "c": 1, "a": 2}) == "[a=2, b=0, c=1]"
-
-
-def test_sorted_key_list():
-    values = {
-        "red": 2,
-        "blue": 3,
-        "orange": 5,
-        "white": 6,
-        "yellow": 7,
-        "green": 9,
-        "black": 0,
-    }
-
-    to_insert = ["red", "blue", "orange", "yellow"]
-
-    skl = SortedKeyList(
-        to_insert,
-        keyfunc=lambda x: values[x],
-        index=list(range(len(to_insert))),
-    )
-
-    assert skl.eq("black") is None
-    assert skl.gte("black").values[0] == "red"
-    assert skl.lte("black") is None
-    skl.insert("black")
-    assert skl.gte("black").values[0] == "black"
-    assert skl.lte("black").values[-1] == "black"
-    assert skl.eq("black").values == ["black"]
-
-    assert skl.gte("white").values[0] == "yellow"
-    assert skl.lte("white").values[-1] == "orange"
-    skl.insert("white")
-    assert skl.gte("white").values[0] == "white"
-    assert skl.gt("white").values[0] == "yellow"
-    assert skl.lte("white").values[-1] == "white"
-    assert skl.lt("yellow").values[-1] == "white"
-
-    assert skl.gte("green") is None
-    assert skl.lte("green").values[-1] == "yellow"
-    skl.insert("green")
-    assert skl.gte("green").values[0] == "green"
-    assert skl.lte("green").values[-1] == "green"
-
-    skl.insert("green")
-    assert skl.eq("green").values == ["green", "green"]
-
-    assert set(skl.ne("green").values) == set(list(values.keys())) - {"green"}
-    values["pokadot"] = -1
-    assert set(skl.ne("pokadot").values) == set(list(values.keys())) - {
-        "pokadot"
-    }
